Pull filterMarkets into separate files
