+++ Sources/FINNFilterImplementation/FilterMarket.swift
@@ -12,239 +12,11 @@ protocol FilterConfiguration {
 }
 
 enum FilterMarket: FilterConfiguration {
-    enum Bap: String, FilterConfiguration, CaseIterable {
-        case bap
-
-        func handlesVerticalId(_ vertical: String) -> Bool {
-            return rawValue == vertical || vertical.hasPrefix(rawValue + "-")
-        }
-
-        var preferenceFilterKeys: [FilterKey] {
-            return [.searchType, .segment, .condition, .published]
-        }
-
-        var supportedFiltersKeys: [FilterKey] {
-            return [
-                .location,
-                .category,
-                .price,
-            ]
-        }
-
-        var mapFilterKey: FilterKey? {
-            return .location
-        }
-    }
-
-    enum Realestate: String, FilterConfiguration, CaseIterable {
-        case homes = "realestate-homes"
-
-        func handlesVerticalId(_ vertical: String) -> Bool {
-            return rawValue == vertical
-        }
-
-        var preferenceFilterKeys: [FilterKey] {
-            return [.published, .isSold, .isNewProperty, .isPrivateBroker]
-        }
-
-        var supportedFiltersKeys: [FilterKey] {
-            return [
-                .location,
-                .price,
-                .priceCollective,
-                .rent,
-                .area,
-                .noOfBedrooms,
-                .constructionYear,
-                .propertyType,
-                .ownershipType,
-                .facilities,
-                .viewing,
-                .floorNavigator,
-                .energyLabel,
-                .plotArea,
-            ]
-        }
-
-        var mapFilterKey: FilterKey? {
-            return .location
-        }
-    }
-
-    enum Car: String, FilterConfiguration, CaseIterable {
-        case norway = "car-norway"
-        case abroad = "car-abroad"
-
-        func handlesVerticalId(_ vertical: String) -> Bool {
-            return rawValue == vertical
-        }
-
-        var preferenceFilterKeys: [FilterKey] {
-            switch self {
-            case .norway:
-                return [.condition, .published, .priceChanged, .dealerSegment]
-            case .abroad:
-                return [.condition, .published, .priceChanged, .dealerSegment]
-            }
-        }
-
-        var supportedFiltersKeys: [FilterKey] {
-            switch self {
-            case .norway:
-                return [
-                    .make,
-                    .salesForm,
-                    .year,
-                    .mileage,
-                    .price,
-                    .location,
-                    .bodyType,
-                    .engineFuel,
-                    .exteriorColour,
-                    .engineEffect,
-                    .numberOfSeats,
-                    .wheelDrive,
-                    .transmission,
-                    .carEquipment,
-                    .warrantyInsurance,
-                    .wheelSets,
-                    .registrationClass,
-                ]
-            case .abroad:
-                return [
-                    .make,
-                    .salesForm,
-                    .year,
-                    .mileage,
-                    .price,
-                    .location,
-                    .bodyType,
-                    .engineFuel,
-                    .exteriorColour,
-                    .engineEffect,
-                    .numberOfSeats,
-                    .wheelDrive,
-                    .transmission,
-                    .carEquipment,
-                    .warrantyInsurance,
-                    .wheelSets,
-                    .registrationClass,
-                ]
-            }
-        }
-
-        var mapFilterKey: FilterKey? {
-            return .location
-        }
-    }
-
-    enum MC: String, FilterConfiguration, CaseIterable {
-        case mc
-        case mopedScooter = "moped-scooter"
-        case snowmobile
-        case atv
-
-        func handlesVerticalId(_ vertical: String) -> Bool {
-            return rawValue == vertical
-        }
-
-        var preferenceFilterKeys: [FilterKey] {
-            switch self {
-            case .mc:
-                return [.published, .segment, .dealerSegment]
-            default:
-                return [.published, .dealerSegment]
-            }
-        }
-
-        var supportedFiltersKeys: [FilterKey] {
-            switch self {
-            case .mc:
-                return [
-                    .location,
-                    .category,
-                    .make,
-                    .price,
-                    .year,
-                    .mileage,
-                    .engineEffect,
-                    .engineVolume,
-                ]
-            case .mopedScooter:
-                return [
-                    .location,
-                    .category,
-                    .make,
-                    .price,
-                    .year,
-                    .mileage,
-                    .engineEffect,
-                    .engineVolume,
-                ]
-            case .snowmobile, .atv:
-                return [
-                    .location,
-                    .make,
-                    .price,
-                    .year,
-                    .mileage,
-                    .engineEffect,
-                    .engineVolume,
-                ]
-            }
-        }
-
-        var mapFilterKey: FilterKey? {
-            return .location
-        }
-    }
-
-    enum Job: String, FilterConfiguration, CaseIterable {
-        case fullTime = "job-full-time"
-        case partTime = "job-part-time"
-        case management = "job-management"
-
-        func handlesVerticalId(_ vertical: String) -> Bool {
-            return rawValue == vertical
-        }
-
-        var preferenceFilterKeys: [FilterKey] {
-            return [.published]
-        }
-
-        var supportedFiltersKeys: [FilterKey] {
-            switch self {
-            case .partTime:
-                return [
-                    .location,
-                    .occupation,
-                    .industry,
-                    .jobDuration,
-                    .jobSector,
-                ]
-            case .fullTime, .management:
-                return [
-                    .location,
-                    .occupation,
-                    .industry,
-                    .jobDuration,
-                    .extent,
-                    .jobSector,
-                    .managerRole,
-                ]
-            }
-        }
-
-        var mapFilterKey: FilterKey? {
-            return .location
-        }
-    }
-
-    case bap(Bap)
-    case realestate(Realestate)
-    case car(Car)
-    case mc(MC)
-    case job(Job)
+    case bap(FilterMarketBap)
+    case realestate(FilterMarketRealestate)
+    case car(FilterMarketCar)
+    case mc(FilterMarketMC)
+    case job(FilterMarketJob)
 
     init?(market: String) {
         guard let market = FilterMarket.allCases.first(where: { $0.handlesVerticalId(market) }) else {
@@ -288,10 +60,10 @@ enum FilterMarket: FilterConfiguration {
 
 extension FilterMarket: CaseIterable {
     static var allCases: [FilterMarket] {
-        return Bap.allCases.map(FilterMarket.bap)
-            + Realestate.allCases.map(FilterMarket.realestate)
-            + Car.allCases.map(FilterMarket.car)
-            + MC.allCases.map(FilterMarket.mc)
-            + Job.allCases.map(FilterMarket.job)
+        return FilterMarketBap.allCases.map(FilterMarket.bap)
+            + FilterMarketRealestate.allCases.map(FilterMarket.realestate)
+            + FilterMarketCar.allCases.map(FilterMarket.car)
+            + FilterMarketMC.allCases.map(FilterMarket.mc)
+            + FilterMarketJob.allCases.map(FilterMarket.job)
     }
 }
